requirements.txt
python-dotenv            # зчитування .env
python-telegram-bot      # робота з Telegram Bot API
schedule                 # простий планувальник
feedparser               # RSS-фіди
beautifulsoup4           # HTML-скрапінг
requests                 # HTTP-запити
transformers             # NLP-моделі (sum, zero-shot)
sentence-transformers    # embeddings для дедуплікації
